Enum server_cli::Command

pub(crate) enum Command {
    Hello {
        high_version: i32,
        low_version: i32,
    },
    CreateSegment {
        segment: String,
        target_rate: i32,
        scale_type: u8,
        delegation_token: String,
    },
    SetupAppend {
        segment: String,
        writer_id: u128,
        delegation_token: String,
    },
    ConditionalAppend {
        segment: String,
        writer_id: u128,
        event_number: i64,
        expected_offset: i64,
        event: String,
        delegation_token: String,
    },
    ReadSegment {
        segment: String,
        offset: i64,
        suggested_length: i32,
        delegation_token: String,
    },
    GetSegmentAttribute {
        segment: String,
        attribute_id: u128,
        delegation_token: String,
    },
    UpdateSegmentAttribute {
        segment: String,
        attribute_id: u128,
        new_value: i64,
        expected_value: i64,
        delegation_token: String,
    },
    GetStreamSegmentInfo {
        segment: String,
        delegation_token: String,
    },
    CreateTableSegment {
        segment: String,
        delegation_token: String,
    },
    UpdateSegmentPolicy {
        segment: String,
        target_rate: i32,
        scale_type: u8,
        delegation_token: String,
    },
    MergeSegments {
        source: String,
        target: String,
        delegation_token: String,
    },
    MergeTableSegments {
        source: String,
        target: String,
        delegation_token: String,
    },
    SealSegment {
        segment: String,
        delegation_token: String,
    },
    SealTableSegment {
        segment: String,
        delegation_token: String,
    },
    TruncateSegment {
        segment: String,
        truncation_offset: i64,
        delegation_token: String,
    },
    DeleteSegment {
        segment: String,
        delegation_token: String,
    },
    DeleteTableSegment {
        segment: String,
        must_be_empty: bool,
        delegation_token: String,
    },
    UpdateTableEntries {
        segment: String,
        key: String,
        key_version: i64,
        value: String,
        delegation_token: String,
        table_segment_offset: i64,
    },
    RemoveTableKeys {
        segment: String,
        table_keys: Vec<String>,
        keys_version: Vec<i64>,
        delegation_token: String,
        table_segment_offset: i64,
    },
    ReadTable {
        segment: String,
        table_keys: Vec<String>,
        keys_version: Vec<i64>,
        delegation_token: String,
    },
    ReadTableKeys {
        segment: String,
        suggested_key_count: i32,
        continuation_token: String,
        delegation_token: String,
    },
    ReadTableEntries {
        segment: String,
        suggested_entry_count: i32,
        continuation_token: String,
        delegation_token: String,
    },
}

Variants

Hello

Send Hello to Server.

Fields of Hello

high_version: i32low_version: i32

CreateSegment

Create a segment with the defined configuration.

Fields of CreateSegment

segment: String

The segment name.

target_rate: i32

Desire rate.

scale_type: u8

ScaleType, 0 = NoScale, 1 = ByRateInKbytesPerSec 2 = ByRateInEventsPerSec

delegation_token: String

The delegation token, default value is null.

SetupAppend

Setup an Append to the given segment.

Fields of SetupAppend

segment: String

The segment name.

writer_id: u128

The writer id.

delegation_token: String

The delegation token, default value is null.

ConditionalAppend

Add an event to a segment. It will create the writer if it does not exist.

Fields of ConditionalAppend

segment: String

The segment name.

writer_id: u128

The writer id.

event_number: i64

The number of the event.

expected_offset: i64

The start offset of the event.

event: String

The content of event.

delegation_token: String

The delegation token, default value is null.

ReadSegment

Read some data from one segment.

Fields of ReadSegment

segment: String

The segment name.

offset: i64

The start offset to read.

suggested_length: i32

The suggested_length of data.

delegation_token: String

The delegation token, default value is null.

GetSegmentAttribute

Get segment attribute of the given segment.

Fields of GetSegmentAttribute

segment: String

The segment name.

attribute_id: u128

The id of the attribute.

delegation_token: String

The delegation token, default value is null.

UpdateSegmentAttribute

Update segment attribute of the given segment.

Fields of UpdateSegmentAttribute

segment: String

The segment name.

attribute_id: u128

The id of the attribute.

new_value: i64

The new value of the attribute.

expected_value: i64

The old value of the attribute, default value is i64::min_value(). which means to add a new attribute.

delegation_token: String

The delegation token, default value is null.

GetStreamSegmentInfo

Get stream segment info.

Fields of GetStreamSegmentInfo

segment: String

The segment name.

delegation_token: String

The delegation token, default value is null.

CreateTableSegment

Create table segment.

Fields of CreateTableSegment

segment: String

The table segment name.

delegation_token: String

The delegation token, default value is null.

UpdateSegmentPolicy

Update segment policy.

Fields of UpdateSegmentPolicy

segment: String

The segment name.

target_rate: i32

Desired rate.

scale_type: u8

ScaleType, 0 = NoScale, 1 = ByRateInKbytesPerSec 2 = ByRateInEventsPerSec

delegation_token: String

The delegation token, default value is null.

MergeSegments

Merge two segments.

Fields of MergeSegments

source: String

The source name of segment.

target: String

The target name of segment.

delegation_token: String

The delegation token, default value is null.

MergeTableSegments

Merge two table segments.

Fields of MergeTableSegments

source: String

The source name of table segment.

target: String

The target name of table segment.

delegation_token: String

The delegation token, default value is null.

SealSegment

Seal segment.

Fields of SealSegment

segment: String

The segment name.

delegation_token: String

The delegation token, default value is null.

SealTableSegment

Seal table segment.

Fields of SealTableSegment

segment: String

The segment name.

delegation_token: String

The delegation token, default value is null.

TruncateSegment

Truncate segment.

Fields of TruncateSegment

segment: String

The segment name.

truncation_offset: i64

The truncation offset.

delegation_token: String

The delegation token, default value is null.

DeleteSegment

Delete segment.

Fields of DeleteSegment

segment: String

The segment name.

delegation_token: String

The delegation token, default value is null.

DeleteTableSegment

Delete table segment.

Fields of DeleteTableSegment

segment: String

The table segment name.

must_be_empty: bool

If true, the segment only allows to delete if it is empty.

delegation_token: String

The delegation token, default value is null.

UpdateTableEntries

Update table entries, for the convenience, it only allows update one entry at one time.

Fields of UpdateTableEntries

segment: String

The segment name.

key: String

The key.

key_version: i64

The key version, default value is i64::min_value(). which means to add a new table entry.

value: String

The value.

delegation_token: String

The delegation token, default value is null.

table_segment_offset: i64

The expected offset to update.

RemoveTableKeys

Remove table keys.

Fields of RemoveTableKeys

segment: String

The segment name.

table_keys: Vec<String>

The data of the table key.

keys_version: Vec<i64>

The version of the table key, the length should be the same as keys array..

delegation_token: String

The delegation token, default value is null.

table_segment_offset: i64

The expected offset to update.

ReadTable

Read table.

Fields of ReadTable

segment: String

The segment name.

table_keys: Vec<String>

The data of the table key.

keys_version: Vec<i64>

The version of the table key, the length should be the same.

delegation_token: String

The delegation token, default value is null.

ReadTableKeys

Read table keys.

Fields of ReadTableKeys

segment: String

The segment name.

suggested_key_count: i32continuation_token: String

This is used to indicate the point from which the next keys should be fetch. use comma to separate the number.

delegation_token: String

The delegation token, default value is null.

ReadTableEntries

Read table entries.

Fields of ReadTableEntries

segment: String

The segment name.

suggested_entry_count: i32continuation_token: String

This is used to indicate the point from which the next keys should be fetch. use comma to separate the number.

delegation_token: String

The delegation token, default value is null.